Merge in DBJAV/dbncontrib_new2 from filter_tables to contrib/aaarrass/nl2sql

* commit 'af3250e69f6c2a4a0500882d698ad53d68da8eea [formerly aaac4b59e757570b6f3eec5708063505549ef390]':
  fix filtering and table selection


Former-commit-id: d67dced0d08a278e5b7bf53ace294c41105aca15

+++ src/main/java/com/dbn/oracleAI/config/ui/profiles/ProfileEditionProviderStep.java
@@ -107,7 +107,7 @@ public class ProfileEditionProviderStep extends WizardStep<ProfileEditionWizardM
   public WizardStep<ProfileEditionWizardModel> onNext(ProfileEditionWizardModel model) {
     profile.setProvider(ProviderType.valueOf(Objects.requireNonNull(providerNameCombo.getSelectedItem()).toString()));
     profile.setModel((ProviderModel) providerModelCombo.getSelectedItem());
-    profile.setTemperature((double) temperatureSlider.getValue() / 10);
+    profile.setTemperature((float) temperatureSlider.getValue() / 10);
     return super.onNext(model);
   }
 }
